Prevent autosave association with has_one defined on child class
Adjusts the logic in inverse_polymorphic_association_changed? to determine
if a has_one association's polymorphic association should autosave the
associated record for a type change.

Previously this would see the type as changed every time if the has_one
is defined on a child class, since the polymorphic relationship saves the
parent class as the *_type value. The new check resolves the correct
class name to avoid these extra saves.

Fixes #51280